Move new things to Library
