+++ Public/Src/Cache/ContentStore/Library/Service/Grpc/GrpcCopyClientKey.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace BuildXL.Cache.ContentStore.Distributed.Utilities
+namespace BuildXL.Cache.ContentStore.Service.Grpc
 {
     /// <summary>
     /// Key to lookup <see cref="GrpcCopyClient"/> in <see cref="GrpcCopyClientCache"/>.
