+++ Public/Src/Cache/ContentStore/Library/Utils/BandwidthChecker.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using BuildXL.Cache.ContentStore.Interfaces.Extensions;
 using BuildXL.Cache.ContentStore.Tracing.Internal;
 
-namespace BuildXL.Cache.ContentStore.Distributed.Utilities
+namespace BuildXL.Cache.ContentStore.Utils
 {
     /// <summary>
     /// Checks that a copy has a minimum bandwidth, and cancells copies otherwise.
