+++ Public/Src/Cache/ContentStore/Test/Utils/ByteArrayComparerTests.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Collections.Generic;
 using BuildXL.Cache.ContentStore.Interfaces.Utils;
 using ContentStoreTest.Test;
 using Xunit;
